Estimation Functions

HighFrequencyCovariance.bnhls_covariance

HighFrequencyCovariance.eigenvalue_clean

HighFrequencyCovariance.estimate_covariance

HighFrequencyCovariance.estimate_microstructure_noise

HighFrequencyCovariance.estimate_volatility

HighFrequencyCovariance.identity_regularisation

HighFrequencyCovariance.nearest_correlation_matrix

HighFrequencyCovariance.nearest_psd_matrix

HighFrequencyCovariance.preaveraged_covariance

HighFrequencyCovariance.regularise

HighFrequencyCovariance.simple_covariance

HighFrequencyCovariance.simple_volatility

HighFrequencyCovariance.spectral_covariance

HighFrequencyCovariance.two_scales_covariance

HighFrequencyCovariance.two_scales_volatility

Estimating Volatility

The estimate_volatility function is the main volatility estimation function. Either of the two estimation methods can be called by specifying :simple_volatility or :two_scales_volatility as the method argument in the estimate_volatility function. Alternatively the simple_volatility or two_scales_volatility functions can be called directly.

The simple_volatility returns a Dict with the estimated volatility for each asset. The two_scales_volatility function on the other hand returns a tuple with a Dict of estimated volatilities in the first position and a Dict of estimated microstructure noise variances in the second. For uniformity of output the estimate_volatility returns a Dict with the estimated volatility for each asset regardless of what method is chosen.

If a user wants to calculate both volatilities and microstructure noises then they are advised to prefer the two_scales_volatility function over doing both estimate_volatility (with the :two_scale_covariance method argument) and the estimate_microstructure_noise function. While the results are the same doing the two function option means everything is calculated twice.

HighFrequencyCovariance.estimate_volatility — Function

This is a convenience wrapper for the two volatility estimation techniques included in this package. The method can be :simplevolatility or :twoscales_volatility in which case the simple or two scales volatilty methods will be called.

estimate_volatility(ts::SortedDataFrame, assets::Vector{Symbol} = get_assets(ts), method::Symbol = :two_scales_volatility;
                    num_grids::Real = default_num_grids(ts),
                    time_grid::Union{Missing,Dict} = missing , fixed_spacing::Union{Missing,Dict,<:Real} = missing,
                    use_all_obs::Bool = false, rough_guess_number_of_intervals::Integer = 5)

HighFrequencyCovariance.simple_volatility — Function

Calculates volatility with the simple method.

simple_volatility(ts::SortedDataFrame, assets::Vector{Symbol} = get_assets(ts);
                  time_grid::Union{Missing,Dict} = missing , fixed_spacing::Union{Missing,Dict,<:Real} = missing,
                  use_all_obs::Bool = false, rough_guess_number_of_intervals::Integer = 5)

Inputs

ts::SortedDataFrame - The tick data.

assets::Vector{Symbol} - The assets you want to estimate volatilities for.

time_grid::Union{Missing,Dict} - The grid with which to calculate returns. If missing one is generated with a fixed spacing (if that is provided) or a default spacing.

fixedspacing::Union{Missing,Dict,<:Real} - A spacing used to calculate a time grid. Not used if a timegrid is input or if useallobs=true.

useallobs::Bool - Use all observations to estimate volatilities. Not used if a time_grid is provided.

roughguessnumberofintervals::Integer - A rough number of intervals to calculate a default spacing. Not used if a timegrid or fixedspacing is provided or if useallobs=true.

T::Real - The duration of the tick data.

Returns

A scalar representing the optimal interval spacing.

HighFrequencyCovariance.two_scales_volatility — Function

Calculates volatility with the two scales method of Zhang, Mykland, Ait-Sahalia 2005. The amount of time for the grid spacing is by default this is a tenth of the total duration by default. If this doesn't make sense for your use of it then choose a spacing at which you expect the effect of microstructure noise will be small.

two_scales_volatility(vals::Vector, times::Vector, asset::Symbol, num_grids::Real)

Inputs

vals::Vector - The prices at each instant in time.

times::Vector - The times corresponding to each element in vals.

asset::Symbol - The name of the asset.

num_grids::Real - Number of grids used in order in two scales estimation.

Returns

A scalar for the estimated volatility of the asset.

A scalar for the estimated microstructure noise variance.

twoscalesvolatility(ts::SortedDataFrame, assets::Vector{Symbol} = getassets(ts); numgrids::Real = defaultnumgrids(ts))

Inputs

ts::SortedDataFrame - The tick data.

assets::Vector{Symbol} - The assets you want to estimate volatilities for.

num_grids::Real - Number of grids used in order in two scales estimation.

Returns

A Dict with estimated volatilities for each asset.

A Dict with estimated microstructure noise variances for each asset.

Estimating Microstructure Noise

There is one function that returns a Dict of microstructure noise estimates for each asset. These estimates come from the two_scales_volatility method and are identical to what you get if you examine the second element of the tuple that that function outputs.

HighFrequencyCovariance.estimate_microstructure_noise — Function

This estimates microstructure noise with the twoscalesvolatility method.

estimate_microstructure_noise(ts::SortedDataFrame, assets::Vector{Symbol} = get_assets(ts);
                              num_grids::Real = default_num_grids(ts),
                              time_grid::Union{Missing,Dict} = missing , fixed_spacing::Union{Missing,Dict,<:Real} = missing,
                              use_all_obs::Bool = false, rough_guess_number_of_intervals::Integer = 5)

Estimating Covariance Matrices

The estimate_covariance is the main method for estimating a CovarianceMatrix. Five possible methods can be input to this function (or the functions for each method can alternatively be called directly).

All covariance estimation functions take in a SortedDataFrame and (optionally) a vector of symbol names representing assets and (optionally) a specified regularisation method. If the vector of symbol names for assets in input then the CovarianceMatrix will only include those input assets and assets will be in the order specified in the vector.

If the regularisation method is specified then this will be used in regularising the resulting matrix. This can alternatively be missing in which case no regularisation will be done. By default the nearest_psd_matrix will be used for every method except the two\_scales\_covariance method and this regularisation is done on the estimated covariance matrix before its correlation matrix and volatilities are split up and placed in a CovarianceMatrix struct. For the two\_scales\_covariance method the correlation matrix is estimated directly and regularisation is applied to this correlation matrix. Hence the nearest_correlation_matrix is the default.

Note that some combinations of estimation technique and regularisation technique will not work. For instance nearest_correlation_matrix would not be good to apply in the case of the preaveraged_covariance method as it would attempt to make a covariance matrix into a correlation matrix with a unit diagonal. In addition if the estimated matrix is very non-psd then heavy regularisation might be required. This may have bad results. In these cases it may be useful to turn off regularisation in the estimation function and instead apply regularisation to the CovarianceMatrix struct.

HighFrequencyCovariance.estimate_covariance — Function

This is a convenience wrapper for the five covariance estimation techniques included in this package. The method can be :simplecovariance, :bnhlscovariance, :spectralcovariance, :preaveragedcovariance or :twoscalescovariance.

estimate_covariance(ts::SortedDataFrame, assets::Vector{Symbol} = get_assets(ts), method::Symbol = :preaveraged_covariance;
                    regularisation::Union{Missing,Symbol} = :default, regularisation_params::Dict = Dict(),
                    only_regulise_if_not_PSD::Bool = false,
                    time_grid::Union{Missing,Vector} = missing,
                    fixed_spacing::Union{Missing,<:Real} = missing, refresh_times::Bool = false, rough_guess_number_of_intervals::Integer = 5, # General Inputs
                    kernel::HFC_Kernel{<:Real} = parzen, H::Real = kernel.c_star * ( mean(map(a -> length(ts.groupingrows[a]), assets))   )^0.6, m::Integer = 2, # BNHLS parameters
                    numJ::Integer = 100, num_blocks::Integer = 10, block_width::Real = (maximum(ts.df[:,ts.time]) - minimum(ts.df[:,ts.time])) / num_blocks, microstructure_noise_var::Dict{Symbol,<:Real} = two_scales_volatility(ts, assets)[2], # Spectral Covariance parameters
                    theta::Real = 0.15, g::NamedTuple = g, # Preaveraging
                    equalweight::Bool = false, num_grids::Real = default_num_grids(ts))

HighFrequencyCovariance.simple_covariance — Function

Estimation of the covariance matrix in the standard simple way. https://en.wikipedia.org/wiki/Samplemeanand_covariance

simple_covariance(ts::SortedDataFrame, assets::Vector{Symbol} = get_assets(ts); regularisation::Union{Missing,Symbol} = :covariance_default, regularisation_params::Dict = Dict(),
                  only_regulise_if_not_PSD::Bool = false, time_grid::Union{Missing,Vector} = missing,
                  fixed_spacing::Union{Missing,<:Real} = missing, refresh_times::Bool = false, rough_guess_number_of_intervals::Integer = 5)

Inputs

ts::SortedDataFrame - The tick data.

assets::Vector{Symbol} - The assets you want to estimate volatilities for.

regularisation::Union{Missing,Symbol} - A symbol representing what regularisation technique should be used. If missing no regularisation is performed.

regularisation_params::Dict - keyword arguments to be consumed in the regularisation algorithm.

onlyreguliseifnotPSD::Bool - Should regularisation only be attempted if the matrix is not psd already.

time_grid::Union{Missing,Dict} - The grid with which to calculate returns.

fixedspacing::Union{Missing,Dict,<:Real} - A spacing used to calculate a time grid. Not used if refreshtimes=true.

refresh_times::Bool - Should refresh times be used to estimate covariance.

roughguessnumberofintervals::Integer - A rough number of intervals to calculate a default spacing. Not used if a timegrid or fixedspacing is provided or if refresh_times=true.

Returns

A CovarianceMatrix.

HighFrequencyCovariance.bnhls_covariance — Function

This calculates covariance with the Multivariate realised kernel oof BNHLS(2011).

bnhls_covariance(ts::SortedDataFrame, assets::Vector{Symbol} = get_assets(ts); regularisation::Union{Missing,Symbol} = :covariance_default, regularisation_params::Dict = Dict(),
                 only_regulise_if_not_PSD::Bool = false, kernel::HFC_Kernel{<:Real} = parzen, H::Real = kernel.c_star * ( mean(map(a -> length(ts.groupingrows[a]), assets))   )^0.6,
                 m::Integer = 2)

Inputs

ts::SortedDataFrame - The tick data.

assets::Vector{Symbol} - The assets you want to estimate volatilities for.

regularisation::Union{Missing,Symbol} - A symbol representing what regularisation technique should be used. If missing no regularisation is performed.

regularisation_params::Dict - keyword arguments to be consumed in the regularisation algorithm.

onlyreguliseifnotPSD::Bool - Should regularisation only be attempted if the matrix is not psd already.

kernel::HFC_Kernel{<:Real} - The kernel used. See the paper for details.

H::Real - The number of lags/leads used in estimation. See the paper for details.

m::Integer - The number of end returns to average.

Returns

A CovarianceMatrix.

References

Barndorff-Nielsen, O., Hansen, P.R., Lunde, A., Shephard, N. 2011. - The whole paper but particularly 2.2, 2.3 here. Kernels are in table 1. choices of H are discussed in section 3.4 of the paper.

HighFrequencyCovariance.spectral_covariance — Function

Estimation of a CovarianceMatrix using the spectral covariance method.

spectral_covariance(ts::SortedDataFrame, assets::Vector{Symbol} = get_assets(ts); regularisation::Union{Missing,Symbol} = :covariance_default, regularisation_params::Dict = Dict(),
                    only_regulise_if_not_PSD::Bool = false, numJ::Integer = 100, num_blocks::Integer = 10, block_width::Real = (maximum(ts.df[:,ts.time]) - minimum(ts.df[:,ts.time])) / num_blocks,
                    microstructure_noise_var::Dict{Symbol,<:Real} = two_scales_volatility(ts, assets)[2])

Inputs

ts::SortedDataFrame - The tick data.

assets::Vector{Symbol} - The assets you want to estimate volatilities for.

regularisation::Union{Missing,Symbol} - A symbol representing what regularisation technique should be used. If missing no regularisation is performed.

regularisation_params::Dict - keyword arguments to be consumed in the regularisation algorithm.

onlyreguliseifnotPSD::Bool - Should regularisation only be attempted if the matrix is not psd already.

numJ::Integer - The number of J values. See the paper for details.

num_blocks::Integer - The number of blocks to split the time frame into. See the paper for details.

block_width::Real - The width of each block to split the time frame into.

microstructurenoisevar::Dict{Symbol,<:Real} - Estimates of microstructure noise variance for each asset.

Returns

A CovarianceMatrix.

References

Bibinger M, Hautsch N, Malec P, Reiss M (2014). “Estimating the quadratic covariation matrix from noisy observations: Local method of moments and efficiency.” The Annals of Statistics, 42(4), 1312–1346. doi:10.1214/14-AOS1224.

HighFrequencyCovariance.preaveraged_covariance — Function

Estimation of the CovarianceMatrix using preaveraging method.

preaveraged_covariance(ts::SortedDataFrame, assets::Vector{Symbol} = get_assets(ts);  regularisation::Union{Missing,Symbol} = :covariance_default, regularisation_params::Dict = Dict(),
                       only_regulise_if_not_PSD::Bool = false, theta::Real = 0.15, g::NamedTuple = g)

Inputs

ts::SortedDataFrame - The tick data.

assets::Vector{Symbol} - The assets you want to estimate volatilities for.

regularisation::Union{Missing,Symbol} - A symbol representing what regularisation technique should be used. If missing no regularisation is performed.

regularisation_params::Dict - keyword arguments to be consumed in the regularisation algorithm.

onlyreguliseifnotPSD::Bool - Should regularisation only be attempted if the matrix is not psd already.

theta::Real - A theta value. See paper for details.

g::NamedTuple - A tuple containing a preaveraging method (with name "f") and a ψ value. See paper for details.

Returns

A CovarianceMatrix.

References

Christensen K, Podolskij M, Vetter M (2013). “On covariation estimation for multivariate continuous Itô semimartingales with noise in non-synchronous observation schemes.” Journal of Multivariate Analysis, 120, 59–84. doi:10.1016/j.jmva.2013.05.002.

HighFrequencyCovariance.two_scales_covariance — Function

Estimation of a CovarianceMatrix using the two scale covariance method.

two_scales_covariance(ts::SortedDataFrame, assets::Vector{Symbol} = get_assets(ts);  regularisation::Union{Missing,Symbol} = :correlation_default, regularisation_params::Dict = Dict(),
                      only_regulise_if_not_PSD::Bool = false, equalweight::Bool = false, num_grids::Real = default_num_grids(ts))

Inputs

ts::SortedDataFrame - The tick data.

assets::Vector{Symbol} - The assets you want to estimate volatilities for.

regularisation::Union{Missing,Symbol} - A symbol representing what regularisation technique should be used. If missing no regularisation is performed.

regularisation_params::Dict - keyword arguments to be consumed in the regularisation algorithm.

onlyreguliseifnotPSD::Bool - Should regularisation only be attempted if the matrix is not psd already.

equalweight::Bool - Should we use equal weight for the two different linear combinations of assets. If false then an optimal weight is calculated (from volatilities).

num_grids::Real - Number of grids used in order in two scales estimation.

Returns

A CovarianceMatrix.

Regularisation of Covariance Matrices

The main function for regularisation is the regularise function. In addition four methods are implemented for regularising matrices can be used directly or through the regularise function. All of these functions can be applied to either a Hermitian matrix or to a CovarianceMatrix struct.

If these functions are applied to a Hermitian then regularisation is applied and a regularised Hermitian is returned.

If these functions are applied to a CovarianceMatrix struct.

HighFrequencyCovariance.regularise — Function

This is a convenience wrapper for the regularisation techniques. The methods can be :identityregularisation, :eigenvalueclean, :nearestcorrelationmatrix or :nearestpsdmatrix. You can also choose :covariancedefault (which is :nearestpsdmatrix) or :correlationdefault (which is :nearestcorrelationmatrix).

regularise(mat::Hermitian, ts::SortedDataFrame,  mat_labels::Vector, method::Symbol = :correlation_default;
           spacing::Union{Missing,<:Real} = missing,
           weighting_matrix = Diagonal(eltype(mat).(I(size(mat)[1]))),
           doDykstra = true, stop_at_first_correlation_matrix = true, max_iterates = 1000)

This is a convenience wrapper for the regularisation techniques. The methods can be :identityregularisation, :eigenvalueclean, :nearestcorrelationmatrix or :nearestpsdmatrix. You can also choose :covariancedefault (which is :nearestpsdmatrix) or :correlationdefault (which is :nearestcorrelationmatrix).

regularise(covariance_matrix::CovarianceMatrix, ts::SortedDataFrame, method::Symbol = :nearest_correlation_matrix;
           spacing::Union{Missing,<:Real} = missing,
           apply_to_covariance::Bool = true,
           weighting_matrix = Diagonal(eltype(covariance_matrix.correlation).(I(size(covariance_matrix.correlation)[1]))),
           doDykstra = true, stop_at_first_correlation_matrix = true, max_iterates = 1000)

HighFrequencyCovariance.identity_regularisation — Function

Regularisation of the correlation matrix by mixing with the identity matrix as per Ledoit & Wolf 2003.

identity_regularisation(mat::Hermitian, asset_returns::DataFrame)

Inputs

mat::Hermitian - A matrix to be regularised.

asset_returns::DataFrame - A DataFrame with returns for each asset.

Returns

A Hermitian.

identityregularisation(mat::Hermitian, assetreturns::DataFrame) where R<:Real

Inputs

mat::Hermitian - A matrix to be regularised.

ts::SortedDataFrame - Tick data.

Returns

A Hermitian.

identityregularisation(covariancematrix::CovarianceMatrix, ts::SortedDataFrame; spacing::Union{Missing,<:Real} = missing, applytocovariance::Bool = true)

This regularises the matrix by doing an elementwise convex linear combination of it with the identity matrix. The weight the identity matrix gets is that specified in Ledoit and Wolf. The inputs are:

covariance_matrix::CovarianceMatrix or mat::Hermitian - The matrix to be regularised.

asset_returns::DataFrame - A DataFrame containing returns for each asset. There should be one column for each asset.

ts::SortedDataFrame - The tick data

spacing::Union{Missing,<:Real} - What spacing (in time) should returns be calculated from ts. If missing refresh times will be used.

applytocovariance::Bool - Should regularisation be applied to the correlation or covariance matrix.

If a Hermitian is input then one will be returned. If a CovarianceMatrix is input then one will be returned.

identity_regularisation(mat::Hermitian, identity_weight::Real)

This regularises the matrix by doing an elementwise convex linear combination of it with the identity matrix (where identity_weight is the weight the identity matrix gets). A Hermitian is returned.

References

Ledoit, O. , Wolf, M. 2003. Improved Estimation of the Covariance Matrix of Stock Returns with an application to portfolio selection. Journal of empirical finance. 10. 603-621.

HighFrequencyCovariance.eigenvalue_clean — Function

Regularisation of the Hermitian matrix by cleaning out small eigenvalues.

eigenvalue_clean(eigenvalues::Vector{<:Real}, eigenvectors::Matrix{<:Real}, eigenvalue_threshold::Real)
eigenvalue_clean(mat::Hermitian, eigenvalue_threshold::Real)

This splits a matrix into its eigenvalues and eigenvectors. Then takes the small eigenvalues (with values below eigenvaluethreshold). It sets them to the greater of their average or eigenvaluethreshold/(4*numberofsmall_eigens). Then the matrix is reconstructed and returned (as a Hermitian)

eigenvalue_clean(mat::Hermitian, ts::SortedDataFrame)
eigenvalue_clean(covariance_matrix::CovarianceMatrix, ts::SortedDataFrame; apply_to_covariance::Bool = true)

Similarly to the above two methods these functions regularise a matrix by setting small eigenvalues to near zero. The method of Laloux, Cizeau, Bouchaud & Potters 2000 is used to choose a threshold. If a Hermitian is input then a Hermitian will be returned. If a CovarianceMatrix is input then one will be returned. If a CovarianceMatrix is input then users can choose to apply regularisation to the covariance matrix or the correlation matrix. If the latter is done then it is advisable to follow up by regularising the correlation matrix with nearest_correlation_matrix to ensure it has a unit diagonal.

Note that if the input matrices include any NaN terms then regularisation is not possible. The matrix will be silently returned (as these NaNs will generally be from upstream problems so it is useful to return the matrix rather than throw at this point).As a result outputs should be checked.

References

Laloux, L., Cizeau, P., Bouchaud J. , Potters, M. 2000. "Random matrix theory and financial correlations" International Journal of Theoretical Applied FInance, 3, 391-397.

HighFrequencyCovariance.nearest_psd_matrix — Function

This function maps a Hermitian matrix to the nearest psd matrix. This uses the projecttoS method in Higham (2001; Theorem 3.2). No special weighting is applied in this case. Advanced users can use the project_to_S directly if they want to use weights in order to decide what the closest pds matrix.

nearest_psd_matrix(mat::Hermitian)
nearest_psd_matrix(covariance_matrix::CovarianceMatrix; apply_to_covariance::Bool = true)
nearest_psd_matrix(covariance_matrix::CovarianceMatrix, ts::SortedDataFrame; apply_to_covariance::Bool = true)

If a Hermitian is input then a Hermitian will be returned. If a CovarianceMatrix is input then a CovarianceMatrix will be returned.

References

Higham NJ (2002). "Computing the nearest correlation matrix - a problem from finance." IMA Journal of Numerical Analysis, 22, 329–343. doi:10.1002/nla.258.

HighFrequencyCovariance.nearest_correlation_matrix — Function

Maps a matrix to the nearest valid correlation matrix (pdf matrix with unit diagonal and all other entries below 1 in absolute value).

nearest_correlation_matrix(covariance_matrix::CovarianceMatrix, ts::SortedDataFrame; weighting_matrix::Union{Diagonal,Hermitian} = Diagonal(eltype(covariance_matrix.correlation).(I(size(covariance_matrix.correlation)[1]))),
                             doDykstra::Bool = true, stop_at_first_correlation_matrix::Bool = true, max_iterates::Integer = 1000)
nearest_correlation_matrix(covariance_matrix::CovarianceMatrix; weighting_matrix::Union{Diagonal,Hermitian} = Diagonal(eltype(covariance_matrix.correlation).(I(size(covariance_matrix.correlation)[1]))),
                             doDykstra::Bool = true, stop_at_first_correlation_matrix::Bool = true, max_iterates::Integer = 1000)
nearest_correlation_matrix(mat::Hermitian, ts::SortedDataFrame, mat_labels = missing; weighting_matrix = Diagonal(eltype(mat).(I(size(mat)[1]))),
                             doDykstra::Bool = true, stop_at_first_correlation_matrix::Bool = true, max_iterates::Integer = 1000)
nearest_correlation_matrix(mat::Hermitian, mat_labels::Vector = missing; weighting_matrix = Diagonal(eltype(mat).(I(size(mat)[1]))),
                             doDykstra::Bool = true, stop_at_first_correlation_matrix::Bool = true, max_iterates::Integer = 1000)

These functions calls the iterate_higham function to move a matrix towards it nearest correlation matrix until it hits a fixed point.

covariance_matrix::CovarianceMatrix or mat::Hermitian - The matrix to be regularised.

ts::SortedDataFrame - The tick data

weighting_matrix::Union{Diagonal,Hermitian} - What weighting matrix should be used (in determining what is the nearest correlation matrix).

doDykstra::Bool Should a Dykstra correction be done.

stopatfirstcorrelationmatrix::Bool Should the iteration stop at the first valid correlation matrix or continue until all iterates have been performed.

maxiterates::Integer What is the maximum number of iterates to do. If stopatfirstcorrelationmatrix = false then maxiterates is the number of iterates that we will do.

If a Hermitian is input then one will be returned. If a CovarianceMatrix is input then one will be returned.

 nearest_correlation_matrix(mat::Union{Diagonal,Hermitian}, W::Union{Diagonal,Hermitian} = Diagonal(Float64.(I(size(mat)[1])));
                             doDykstra::Bool = true, stop_at_first_correlation_matrix::Bool = true, max_iterates::Integer = 1000)

This also iterates a matrix towards convergence but returns a tuple with the updated Hermitian/Diagonal matrix in the first slot, the number of iterates in the second and a Symbol representing a convergence status in the third. The arguments are as above.