Helper Functions

HighFrequencyCovariance.bnhls_2008

HighFrequencyCovariance.fejer

HighFrequencyCovariance.parzen

HighFrequencyCovariance.quadratic_spectral

HighFrequencyCovariance.tukey_hanning

HighFrequencyCovariance.HFC_Kernel

StochasticIntegrals.ItoSet

HighFrequencyCovariance.blockwise_estimation

HighFrequencyCovariance.calculate_mean_abs_distance

HighFrequencyCovariance.dataframe_to_covariancematrix

HighFrequencyCovariance.generate_random_path

HighFrequencyCovariance.get_all_refresh_times

HighFrequencyCovariance.iterate_higham

HighFrequencyCovariance.latest_value

HighFrequencyCovariance.project_to_S

HighFrequencyCovariance.project_to_U

HighFrequencyCovariance.put_assets_into_blocks_by_trading_frequency

HighFrequencyCovariance.random_value_in_interval

HighFrequencyCovariance.squared_frobenius

HighFrequencyCovariance.squared_frobenius_distance

HighFrequencyCovariance.time_between_refreshes

StochasticIntegrals.to_dataframe

Working with SortedDataFrame structs

HighFrequencyCovariance.bnhls_2008

HighFrequencyCovariance.fejer

HighFrequencyCovariance.parzen

HighFrequencyCovariance.quadratic_spectral

HighFrequencyCovariance.tukey_hanning

HighFrequencyCovariance.CovarianceMatrix

HighFrequencyCovariance.HFC_Kernel

HighFrequencyCovariance.SortedDataFrame

StochasticIntegrals.ItoSet

HighFrequencyCovariance.blockwise_estimation

HighFrequencyCovariance.bnhls_covariance

HighFrequencyCovariance.calculate_mean_abs_distance

HighFrequencyCovariance.dataframe_to_covariancematrix

HighFrequencyCovariance.eigenvalue_clean

HighFrequencyCovariance.estimate_covariance

HighFrequencyCovariance.estimate_microstructure_noise

HighFrequencyCovariance.estimate_volatility

HighFrequencyCovariance.generate_random_path

HighFrequencyCovariance.get_all_refresh_times

HighFrequencyCovariance.identity_regularisation

HighFrequencyCovariance.iterate_higham

HighFrequencyCovariance.latest_value

HighFrequencyCovariance.nearest_correlation_matrix

HighFrequencyCovariance.nearest_psd_matrix

HighFrequencyCovariance.preaveraged_covariance

HighFrequencyCovariance.project_to_S

HighFrequencyCovariance.project_to_U

HighFrequencyCovariance.put_assets_into_blocks_by_trading_frequency

HighFrequencyCovariance.random_value_in_interval

HighFrequencyCovariance.regularise

HighFrequencyCovariance.simple_covariance

HighFrequencyCovariance.simple_volatility

HighFrequencyCovariance.spectral_covariance

HighFrequencyCovariance.squared_frobenius

HighFrequencyCovariance.squared_frobenius_distance

HighFrequencyCovariance.time_between_refreshes

HighFrequencyCovariance.two_scales_covariance

HighFrequencyCovariance.two_scales_volatility

StochasticIntegrals.to_dataframe

Working with CovarianceMatrix structs

HighFrequencyCovariance.bnhls_2008

HighFrequencyCovariance.fejer

HighFrequencyCovariance.parzen

HighFrequencyCovariance.quadratic_spectral

HighFrequencyCovariance.tukey_hanning

HighFrequencyCovariance.CovarianceMatrix

HighFrequencyCovariance.HFC_Kernel

HighFrequencyCovariance.SortedDataFrame

StochasticIntegrals.ItoSet

HighFrequencyCovariance.blockwise_estimation

HighFrequencyCovariance.bnhls_covariance

HighFrequencyCovariance.calculate_mean_abs_distance

HighFrequencyCovariance.dataframe_to_covariancematrix

HighFrequencyCovariance.eigenvalue_clean

HighFrequencyCovariance.estimate_covariance

HighFrequencyCovariance.estimate_microstructure_noise

HighFrequencyCovariance.estimate_volatility

HighFrequencyCovariance.generate_random_path

HighFrequencyCovariance.get_all_refresh_times

HighFrequencyCovariance.identity_regularisation

HighFrequencyCovariance.iterate_higham

HighFrequencyCovariance.latest_value

HighFrequencyCovariance.nearest_correlation_matrix

HighFrequencyCovariance.nearest_psd_matrix

HighFrequencyCovariance.preaveraged_covariance

HighFrequencyCovariance.project_to_S

HighFrequencyCovariance.project_to_U

HighFrequencyCovariance.put_assets_into_blocks_by_trading_frequency

HighFrequencyCovariance.random_value_in_interval

HighFrequencyCovariance.regularise

HighFrequencyCovariance.simple_covariance

HighFrequencyCovariance.simple_volatility

HighFrequencyCovariance.spectral_covariance

HighFrequencyCovariance.squared_frobenius

HighFrequencyCovariance.squared_frobenius_distance

HighFrequencyCovariance.time_between_refreshes

HighFrequencyCovariance.two_scales_covariance

HighFrequencyCovariance.two_scales_volatility

StochasticIntegrals.to_dataframe

Blocking and Regularisation Functions

HighFrequencyCovariance.put_assets_into_blocks_by_trading_frequency — Function

This makes a dataframe that describes how to estimate the covariance matrix blockwise.

Takes

ts::SortedDataFrame - The tick data.

obsmultiplefornewblock::Real - The relative number of ticks to seperate a new block. For instance 1.2 means make a new block when some asset has 20% more observations than another in that block.

func::Symbol - A symbol representing the covariance estimation function to be used.

optional_parameters::NamedTuple - Optional parameters to be used in the func function.

Returns

A DataFrame representing what estimations should be performed. The order of rows in the DataFrame shows the order of estimation.

References

Hautsch, N., Kyj, L.M. and Oomen, R.C.A. (2012), A blocking and regularization approach to high‐dimensional realized covariance estimation. J. Appl. Econ., 27: 625-645

HighFrequencyCovariance.blockwise_estimation — Function

Run a series of covariance estimations and combine the results. Two things should be input, a SortedDataFrame with the price update data and a dataframe describing what estimations should be performed. This should be of the same form as is output by putassetsintoblocksbytradingfrequency (although the actual estimations can be customised to something different as to what that function outputs).

Takes

ts::SortedDataFrame - The tick data.

blockingframe::DataFrame - A DataFrame representing what estimations to do and in what order. This is often be one generated by the `putassetsintoblocksbytrading_frequency` function (and potentially then modified).

Returns

A CovarianceMatrix.

Monte Carlo

HighFrequencyCovariance.generate_random_path — Function

Generate a random path of price updates with a specified number of dimensions and ticks. There are options for whether the data is syncronous or asyncronous, the volatility of the price processes, the refresh rate on the (exponential) arrival times of price updates, the minimum and the maximum microstructure noises.

Takes

dimensions::Integer - The number of assets

ticks::Integer - The number of ticks to produce

syncronous::Bool - Should ticks be syncronous (for each asset) or asyncronous

twister::MersenneTwister - The MersenneTwister used for RNG.

minvol::Real - The minimum volatility in sampling (only used if vols is missing)

maxvol::Real - The maximum volatility in sampling (only used if vols is missing)

minrefreshrate::Real - The minimum refresh rate in sampling

maxrefreshrate::Real - The maximum refresh rate in sampling

minnoisevar::Real - The minimum assetwise microstructure noise variance

maxnoisevar::Real - The minimum assetwise microstructure noise variance

assets::Union{Vector,Missing} = missing

browniancorrmatrix::Union{Hermitian,Missing} - The correlation matrix to use. This is sampled from the Inverse Wishart distribution if none is input.

vols::Union{Vector,Missing} - The correlation matrix to use. This is sampled from the Inverse Wishart distribution if none is input.

Returns

A SortedDataFrame of tick data

A CovarianceMatrix representing the true data generation process used in making the tick data

A dict of microstructure_noises for each asset.

A dict of update_rates for each asset.

StochasticIntegrals.ItoSet — Type

Convert a CovarianceMatrix into an ItoSet from the StochasticIntegrals package. This package can then be used to do things like generate draws from the Multivariate Gaussian corresponding to the covariance matrix and other things.

Takes

covariance_matrix::CovarianceMatrix{<:Real}

Returns

A StochasticIntegrals.ItoSet struct.

For getting a DataFrame version of a CovarianceMatrix and vice versa.

StochasticIntegrals.to_dataframe — Function

Convert a CovarianceMatrix to a dataframe format.

Takes

covar::CovarianceMatrix

othercols::Dict - This adds columns to the DataFrame. For instance if it is Dict{Symbol,String}([:pc] .=> ["Fred's PC"]), then there will be a column indicating that this estimation was done on Fred's PC.

deleteduplicatecorrelations::Bool - Should the unnecessary correlations be included (as correlation matrices are symmetric half the entries duplicate information).

Returns

A DataFrame.

HighFrequencyCovariance.dataframe_to_covariancematrix — Function

Convert a CovarianceMatrix to a dataframe format.

Takes

dd::DataFrame - A DataFrame of the form generated by the to_dataframe function.

Returns

A CovarianceMatrix struct.

Blocking and Regularisation Functions

Missing docstring.

Missing docstring for put_assets_into_blocks_by_trading_frequency. Check Documenter's build log for details.

Missing docstring.

Missing docstring for blockwise_estimation. Check Documenter's build log for details.

Internal Functions potentially of use for advanced users

Metrics for distances between CovarianceMatrix structs

HighFrequencyCovariance.calculate_mean_abs_distance — Function

calculate_mean_abs_distance(cov1::CovarianceMatrix, cov2::CovarianceMatrix)

Calculates the mean absolute distance (elementwise in L1 norm) between two CovarianceMatrixs. Undefined if any labels differ between the two CovarianceMatrixs.

Takes

cov1::CovarianceMatrix - The first CovarianceMatrix

cov2::CovarianceMatrix - The second CovarianceMatrix

Returns

An Tuple with the distance for correlations in first entry and distance for volatilities in the second.

calculatemeanabs_distance(d1::Dict{Symbol,<:Real}, d2::Dict{Symbol,<:Real})

Calculates the mean absolute distance (elementwise in L1 norm) between two CovarianceMatrixs.

Takes

d1::Dict{Symbol,<:Real} - The first Dict

d2::Dict{Symbol,<:Real} - The second Dict

Returns

A scalar with the mean distance between matching elements.

calculatemeanabs_distance(d1::Tuple{Dict{Symbol,<:Real},Dict{Symbol,<:Real}}, d2::Tuple{Dict{Symbol,<:Real},Dict{Symbol,<:Real}})

Calculates the distance between the first element of d1 and first of d2. And second elements of each. Returns the Tuple.

Takes

d1::Tuple{Dict{Symbol,<:Real},Dict{Symbol,<:Real}}

d2::Tuple{Dict{Symbol,<:Real},Dict{Symbol,<:Real}}

Returns

A Tuple with two scalars.

HighFrequencyCovariance.squared_frobenius — Function

Rearrange the squared frobenius norm of a matrix. Returns a real.

HighFrequencyCovariance.squared_frobenius_distance — Function

Rearrange the squared frobenius distance between two matrices. Returns a real

Used in volatility estimation techniques

Missing docstring.

Missing docstring for default_num_grids. Check Documenter's build log for details.

Used in covariance estimation techniques

HighFrequencyCovariance.get_all_refresh_times — Function

Get a vector of all refresh times when all assets have an updated price. So if there are assets A and B that trade at times (1,5,6,7,10) and (2,5,7,9) then the refresh times are (2,5,7,10) as at these four times there are updated prices for all assets that have happened since the previous refresh time.

Takes

ts::SortedDataFrame - The tick data.

assets::Vector{Symbol} - The assets of interest.

start_time::Real - From what time should we start looking for updated prices.

Returns

A Vector of refresh times.

HighFrequencyCovariance.latest_value — Function

Get the latest price at a each input time.

Takes

ts::SortedDataFrame - The tick data.

at_times::Vector{R} - The times you want the latest prices for.

assets::Vector{Symbol} - The assets you want latest prices for

Returns

A DataFrame. Rows are for each time specified in at_times. Columns are for each asset.

HighFrequencyCovariance.time_between_refreshes — Function

Get a DataFrame showing how many time is between each refresh and how many ticks in total.

Takes

ts::SortedDataFrame - Tick data.

assets::Vector{Symbol} - A Vector of labels.

Returns

A DataFrame summarising the average number of time between ticks for each asset.

HighFrequencyCovariance.random_value_in_interval — Function

Get a random value in an interval. So if you input times 1,7,8 then for the second entry it will pick a random update (if any exist) between times 1 and 7.

Takes

ts::SortedDataFrame - The tick data.

at_times::Vector{R} - The times that seperate the intervals of interest.

assets::Vector{Symbol} - The assets of interest.

twisterarbvalueininterval::MersenneTwister - The RNG used in selecting the random interval.

Returns

A DataFrame with prices for each asset from random ticks in each interval.

Kernels used in the BNHLS method

HighFrequencyCovariance.HFC_Kernel — Type

A kernel used in the bnhls covariance method.

HighFrequencyCovariance.parzen — Constant

A parzen kernel used in the bnhls covariance method.

HighFrequencyCovariance.quadratic_spectral — Constant

A quadratic_spectral kernel used in the bnhls covariance method.

HighFrequencyCovariance.fejer — Constant

A fejer kernel used in the bnhls covariance method.

HighFrequencyCovariance.tukey_hanning — Constant

A tukey_hanning kernel used in the bnhls covariance method.

HighFrequencyCovariance.bnhls_2008 — Constant

A bnhls_2008 kernel used in the bnhls covariance method.

Used in nearest correlation regularisation

HighFrequencyCovariance.project_to_S — Function

This maps a matrix to the nearest psd matrix. Wroot should be the principal square root of a psd Hermitian weighting matrix, W. Winv_sqrt should be the corresponding square root of the inverse of W. Higham, N. J. 2001. Theorem 3.2

nearest_psd_matrix is a simpler interface for this function however it does not allow weighting matrices to be specified.

HighFrequencyCovariance.project_to_U — Function

This maps the Hermitian/Hermitian matrix A to the nearest matrix in the U space (the space of all unit diagonal matrices as defined by Higham 2001). The inverse weight matrix invW determines how much to adjust each element to get it to be unit diagonal. The weight matrix must be Hermitian positive definite. We use the W-norm (as defined by Higham 2001).

References

Higham, N. J. 2001. Bottom of page 335.

HighFrequencyCovariance.iterate_higham — Function

Do one iterate mapping the input matrix to the S space (of psd matrices) and then to the U space (unit diagonal and all other entries below 1 in absolute value). Returns the updated matrix and the next iterate's Dykstra correction. Higham, N. J. 2001. Algorithm 3.3