Structs

HighFrequencyCovariance.CovarianceMatrix

HighFrequencyCovariance.SortedDataFrame

The SortedDataFrame struct wraps a DataFrame (from the DataFrames package). In the constructor function for the DataFrame we presort the data and create a mapping Dict so that it is fast to subset the DataFrame by the group.

The CovarianceMatrix mutable struct stores three elements. A Hermitian correlation matrix, a vector of volatilities and a vector of labels. The order of the labels matches the order of the assets in the volatility vector and correlation matrix.

Structs specification

HighFrequencyCovariance.SortedDataFrame — Type

SortedDataFrame(df::DataFrame, time::Symbol = :Time, grouping::Symbol = :Name, value::Symbol = :Value)

This struct wraps a DataFrame. In the constructor function for the dataframe we presort the data and create a mapping dict so that it is fast to subset the DataFrame by the group.

For the constructor pass in the dataframe, name of time column, name of grouping column and name of value column to the constructor.

Inputs

df - The tick data

time - The column of the data representing time.

grouping - The column of the data representing the asset name

value - The column of the data representing price/logprice/etc.

time_period_per_unit - The period that one unit (in the time column) corresponds to.

Returns

A SortedDataFrame.

HighFrequencyCovariance.CovarianceMatrix — Type

CovarianceMatrix(correlation::Hermitian{R}, volatility::Vector{R},
                 labels::Vector{Symbol}) where R<:Real

This Struct stores three elements. A Hermitian correlation matrix, a vector of volatilities and a vector of labels. The order of the labels matches the order of the assets in the volatility vector and correlation matrix. The default constructor is used.

Inputs

correlation - A Hermitian correlation matrix.

volatility - Volatilities for each asset.

labels - The labels for the correlation and volatility members. The n'th entry of the labels vector should contain the name of the asset that has its volatility in the n'th entry of the volatility member and its correlations in the n'th row/column of the correlation member.

time_period_per_unit - The period that one unit of volatility corresponds to.

Returns

A CovarianceMatrix.