Functions

HighFrequencyCovariance.bnhls_covariance

HighFrequencyCovariance.preaveraged_covariance

HighFrequencyCovariance.simple_covariance

HighFrequencyCovariance.spectral_covariance

HighFrequencyCovariance.two_scales_covariance

Estimating Covariance Matrices

HighFrequencyCovariance.simple_covariance — Function

     simple_covariance(ts::SortedDataFrame, assets::Vector{Symbol} = get_assets(ts); regularisation::Union{Missing,Symbol} = :covariance_default, regularisation_params::Dict = Dict(),
                       only_regulise_if_not_PSD::Bool = false, time_grid::Union{Missing,Vector} = missing,
                       fixed_spacing::Union{Missing,<:Real} = missing, refresh_times::Bool = false, rough_guess_number_of_intervals::Integer = 5)

Estimation of the covariance matrix in the standard simple way. https://en.wikipedia.org/wiki/Samplemeanand_covariance

HighFrequencyCovariance.bnhls_covariance — Function

This calculates covariance with the Multivariate realised kernel oof BNHLS(2011) Barndorff-Nielsen, O., Hansen, P.R., Lunde, A., Shephard, N. 2011. - The whole paper but particularly 2.2, 2.3 here. Kernels are in table 1. choices of H are discussed in section 3.4 of the paper.

HighFrequencyCovariance.spectral_covariance — Function

Estimation of a CovarianceMatrix using the spectral covariance method. Bibinger M, Hautsch N, Malec P, Reiss M (2014). “Estimating the quadratic covariation matrix from noisy observations: Local method of moments and efficiency.” The Annals of Statistics, 42(4), 1312–1346. doi:10.1214/14-AOS1224.

HighFrequencyCovariance.preaveraged_covariance — Function

Estimation of the CovarianceMatrix using preaveraging method. Christensen K, Podolskij M, Vetter M (2013). “On covariation estimation for multivariate continuous Itô semimartingales with noise in non-synchronous observation schemes.” Journal of Multivariate Analysis, 120, 59–84. doi:10.1016/j.jmva.2013.05.002.

HighFrequencyCovariance.two_scales_covariance — Function

Estimation of a CovarianceMatrix using the two scale covariance method.