Helper Functions

HighFrequencyCovariance.bnhls_2008

HighFrequencyCovariance.fejer

HighFrequencyCovariance.parzen

HighFrequencyCovariance.quadratic_spectral

HighFrequencyCovariance.tukey_hanning

HighFrequencyCovariance.HFC_Kernel

StochasticIntegrals.ItoSet

HighFrequencyCovariance.blockwise_estimation

HighFrequencyCovariance.calculate_mean_abs_distance

HighFrequencyCovariance.combine_covariance_matrices

HighFrequencyCovariance.construct_matrix_from_eigen

HighFrequencyCovariance.cor2cov

HighFrequencyCovariance.cov2cor

HighFrequencyCovariance.cov2cor_and_vol

HighFrequencyCovariance.covariance

HighFrequencyCovariance.dataframe_to_covariancematrix

HighFrequencyCovariance.default_num_grids

HighFrequencyCovariance.duration

HighFrequencyCovariance.generate_random_path

HighFrequencyCovariance.get_all_refresh_times

HighFrequencyCovariance.get_assets

HighFrequencyCovariance.get_correlation

HighFrequencyCovariance.get_returns

HighFrequencyCovariance.get_volatility

HighFrequencyCovariance.is_psd_matrix

HighFrequencyCovariance.iterate_higham

HighFrequencyCovariance.latest_value

HighFrequencyCovariance.make_nan_covariance_matrix

HighFrequencyCovariance.project_to_S

HighFrequencyCovariance.project_to_U

HighFrequencyCovariance.put_assets_into_blocks_by_trading_frequency

HighFrequencyCovariance.random_value_in_interval

HighFrequencyCovariance.rearrange

HighFrequencyCovariance.squared_frobenius

HighFrequencyCovariance.squared_frobenius_distance

HighFrequencyCovariance.subset_to_tick

HighFrequencyCovariance.subset_to_time

HighFrequencyCovariance.ticks_per_asset

HighFrequencyCovariance.time_between_refreshes

HighFrequencyCovariance.valid_correlation_matrix

StochasticIntegrals.to_dataframe

Working with SortedDataFrame structs

HighFrequencyCovariance.get_assets — Function

get_assets(ts::SortedDataFrame, obs_to_include::Integer = 10)

This returns a vector of all of the assets in the SortedDataFrame with at least some number of observations (10 by default).

Inputs

ts - The tick data.

obs_to_include - An integer for the minimum number of ticks in ts needed for the function to include that asset.

Returns

A Vector{Symbol} with each asset.

HighFrequencyCovariance.ticks_per_asset — Function

ticks_per_asset(ts::SortedDataFrame, assets::Vector{Symbol} = get_assets(ts))

Count the number of observations for each asset.

Inputs

ts - The tick data

assets - A vector with asset Symbols.

Returns

A Dict with the number of observations for each input asset.

HighFrequencyCovariance.duration — Function

duration(ts::SortedDataFrame)

The time elapsed between the first and the last tick in a SortedDataFrame.

Inputs

ts - Tick data.

Returns

A scalar representing this duration.

HighFrequencyCovariance.subset_to_tick — Function

subset_to_tick(ts::SortedDataFrame, n::Integer)

This subsets a SortedDataFrame to only the first n ticks.

Inputs

ts - Tick data.

n - How many ticks to subset to.

Returns

A (smaller) SortedDataFrame.

HighFrequencyCovariance.subset_to_time — Function

subset_to_time(ts::SortedDataFrame, totime::Real)

This subsets a SortedDataFrame to only the first observations up until some time.

Inputs

ts - Tick data.

totime - Up to what time.

Returns

A (smaller) SortedDataFrame.

Working with CovarianceMatrix structs

HighFrequencyCovariance.covariance — Function

covariance(cm::CovarianceMatrix, duration::Real)

This makes a Hermitian matrix for the covariance matrix over some duration.

Inputs

cm - A CovarianceMatrix struct.

duration - A duration. This should be in same units as used in estimating cm's volatilities.

Returns

A Hermitian.

HighFrequencyCovariance.get_correlation — Function

get_correlation(covar::CovarianceMatrix, asset1::Symbol, asset2::Symbol)

Extract the correlation between two assets stored in a CovarianceMatrix.

Inputs

covar - A CovarianceMatrix

asset1 - A Symbol representing an asset.

asset2 - A Symbol representing an asset.

Returns

A Scalar (the correlation coefficient).

HighFrequencyCovariance.get_volatility — Function

get_volatility(covar::CovarianceMatrix, asset1::Symbol)

Get the volatility for a stock from a CovarianceMatrix.

Inputs

covar - A CovarianceMatrix

asset1 - A Symbol representing an asset.

Returns

A Scalar (the volatility).

HighFrequencyCovariance.make_nan_covariance_matrix — Function

make_nan_covariance_matrix(labels::Vector{Symbol})

This makes an empty CovarianceMatrix struct with all volatilities and correlations being NaNs.

Inputs

labels - The names of the asset names for this (empty) CovarianceMatrix.

Returns

An (empty) CovarianceMatrix

HighFrequencyCovariance.combine_covariance_matrices — Function

combine_covariance_matrices(vect::Vector{CovarianceMatrix{REAL}}, cor_weights::Vector{<:Real} = repeat([1.0], length(vect)),
                            vol_weights::Vector{<:Real} = cor_weights) where REAL<:Real

Combines a vector of CovarianceMatrix structs into one CovarianceMatrix struct.

Inputs

vect - A vector of CovarianceMatrix structs.

cor_weights - A vector for how much to weight the correlations from each covariance matrix (by default they will be equalweighted).

vol_weights - A vector for how much to weight the volatilities from each covariance matrix (by default they will be equalweighted).

Returns

A matrix and a vector of labels for each row/column of the matrix.

HighFrequencyCovariance.rearrange — Function

rearrange(cm::CovarianceMatrix, labels::Vector{Symbol})

Rearrange the order of labels in a CovarianceMatrix.

Takes

cm - A CovarianceMatrix.

labels - A Vector of labels.

Returns

A CovarianceMatrix.

HighFrequencyCovariance.cov2cor — Function

cov2cor(mat::AbstractMatrix)

Converts a matrix (representing a covariance matrix) into a Hermitian correlation matrix and a vector of standard deviations.

Inputs

cor - A matrix.

Returns

A Hermitian.

A Vector of standard deviations (not volatilities).

HighFrequencyCovariance.cor2cov — Function

cor2cov(cor::AbstractMatrix,sdevs::Vector{<:Real})

Converts a correlation matrix and some standard deviations into a Hermitian covariance matrix.

Inputs

cor - A correlation matrix.

sdevs - A vector of standard deviations (not volatilities - use sdevs = sqrt(duration) .* volatilities to convert if necessary).

Returns

A Hermitian.

HighFrequencyCovariance.cov2cor_and_vol — Function

cov2cor_and_vol(mat::AbstractMatrix, duration::Real)

Converts a matrix (representing a covariance matrix) into a Hermitian correlation matrix and a vector of volatilities.

Inputs

cor - A correlation matrix.

duration - The duration that the covariance matrix is for.

Returns

A Hermitian.

A Vector of volatilities.

HighFrequencyCovariance.construct_matrix_from_eigen — Function

construct_matrix_from_eigen(eigenvalues::Vector{<:Real}, eigenvectors::Matrix{<:Real})

Constructs a matrix from its eigenvalue decomposition.

Inputs

eigenvalues - A vector of eigenvalues.

eigenvectors - A matrix of eigenvectors. The i'th column corresponds to the i'th eigenvalue.

Returns

A Matrix.

HighFrequencyCovariance.get_returns — Function

get_returns(dd::DataFrame; rescale_for_duration::Bool = false)

Converts a long format DataFrame of prices into a DataFrame of returns.

Inputs

dd - A DataFrame with a column called :Time and all other columns being asset prices in each period.

rescale_for_duration - Should returns be rescaled to reflect a common time interval.

Returns

A DataFrame of returns.

HighFrequencyCovariance.valid_correlation_matrix — Function

valid_correlation_matrix(mat::Hermitian)
valid_correlation_matrix(covar::CovarianceMatrix)

Test if a Hermitian matrix is a valid correlation matrix. This is done by testing if it is psd, if it has a unit diagonal and if all other elements are less than one. If a Hermitian is input then it will be tested. If a CovarianceMatrix is input then its correlation matrix will be tested.

Inputs

A Hermitian matrix or a CovarianceMatrix

Returns

A Bool that is true if mat is a valid correlation matrix and false if not.

HighFrequencyCovariance.is_psd_matrix — Function

is_psd_matrix(mat::Hermitian)
is_psd_matrix(covar::CovarianceMatrix)

Test if a matrix is psd (Positive Semi-Definite). This is done by seeing if all eigenvalues are positive. If a Hermitian is input then it will be tested. If a CovarianceMatrix is input then its correlation matrix will be tested.

Inputs

A Hermitian matrix or a CovarianceMatrix

Returns

A Bool that is true if mat is psd and false if not.

Blocking and Regularisation Functions

HighFrequencyCovariance.put_assets_into_blocks_by_trading_frequency — Function

put_assets_into_blocks_by_trading_frequency(ts::SortedDataFrame, obs_multiple_for_new_block::Real, func::Symbol, optional_parameters::NamedTuple = NamedTuple())

This makes a DataFrame that describes how to estimate the covariance matrix blockwise.

Inputs

ts - The tick data.

obs_multiple_for_new_block - The relative number of ticks needed before a new block is made. So if this is 1.2 that means a new group is made when one asset has 20% or more ticks than the slowest traded asset in the previous block.

func - A symbol representing the covariance estimation function to be used.

optional_parameters - Optional parameters to be used in the func function.

Returns

A DataFrame representing what estimations should be performed. The order of rows in the DataFrame shows the order of estimation.

References

Hautsch, N., Kyj, L.M. and Oomen, R.C.A. (2012), A blocking and regularization approach to high‐dimensional realized covariance estimation. J. Appl. Econ., 27: 625-645

HighFrequencyCovariance.blockwise_estimation — Function

blockwise_estimation(ts::SortedDataFrame, blocking_frame::DataFrame)

Run a series of covariance estimations and combine the results. Two things should be input, a SortedDataFrame with the price update data and a dataframe describing what estimations should be performed. This should be of the same form as is output by put_assets_into_blocks_by_trading_frequency (although the actual estimations can be customised to something different as to what that function outputs).

Inputs

ts - The tick data.

blocking_frame - A DataFrame representing what estimations to do and in what order. This is often be one generated by the put_assets_into_blocks_by_trading_frequency function (and potentially then modified).

Returns

A CovarianceMatrix.

Monte Carlo

HighFrequencyCovariance.generate_random_path — Function

generate_random_path(dimensions::Integer, ticks::Integer; syncronous::Bool = false, twister::MersenneTwister = MersenneTwister(1), minvol::Real = 0.0, maxvol::Real = 0.02,
                     min_refresh_rate::Real = 1.0, max_refresh_rate::Real = 5.0, min_noise_var::Real = 0.0, max_noise_var::Real = 0.01, assets::Union{Vector,Missing} = missing,
                     brownian_corr_matrix::Union{Hermitian,Missing} = missing, vols::Union{Vector,Missing} = missing)

Generate a random path of price updates with a specified number of dimensions and ticks. There are options for whether the data is syncronous or asyncronous, the volatility of the price processes, the refresh rate on the (exponential) arrival times of price updates, the minimum and the maximum microstructure noises.

Inputs

dimensions - The number of assets.

ticks - The number of ticks to produce.

syncronous - Should ticks be syncronous (for each asset) or asyncronous.

twister - The MersenneTwister used for RNG.

minvol - The minimum volatility in sampling (only used if vols is missing).

maxvol - The maximum volatility in sampling (only used if vols is missing).

min_refresh_rate - The minimum refresh rate in sampling.

max_refresh_rate - The maximum refresh rate in sampling.

min_noise_var - The minimum assetwise microstructure noise variance.

max_noise_var - The minimum assetwise microstructure noise variance.

assets - The names of the assets that you want to use. The length of this must be equal to the dimensions input.

brownian_corr_matrix - The correlation matrix to use. This is sampled from the Inverse Wishart distribution if none is input.

vols - The correlation matrix to use. This is sampled from the Inverse Wishart distribution if none is input.

Returns

A SortedDataFrame of tick data.

A CovarianceMatrix representing the true data generation process used in making the tick data.

A Dict of microstructure noise variances for each asset.

A Dict of update rates for each asset.

StochasticIntegrals.ItoSet — Type

ItoSet(covariance_matrix::CovarianceMatrix{<:Real})

Convert a CovarianceMatrix into an ItoSet from the StochasticIntegrals package. This package can then be used to do things like generate draws from the Multivariate Gaussian corresponding to the covariance matrix and other things.

Inputs

covariance_matrix - The CovarianceMatrix that you want to convert into an StochasticIntegrals.ItoSet

Returns

A StochasticIntegrals.ItoSet struct.

For getting a DataFrame version of a CovarianceMatrix and vice versa.

StochasticIntegrals.to_dataframe — Function

to_dataframe(covar::CovarianceMatrix, othercols::Dict = Dict{Symbol,Any}(); delete_duplicate_correlations::Bool = true)

Convert a CovarianceMatrix to a DataFrame format.

Inputs

covar - The CovarianceMatrix

othercols - This adds columns to the DataFrame. For instance if it is Dict{Symbol,String}([:pc] .=> ["Fred's PC"]), then there will be a column indicating that this estimation was done on Fred's PC.

delete_duplicate_correlations - Should the unnecessary correlations be included (as correlation matrices are symmetric half the entries duplicate information).

Returns

A DataFrame.

HighFrequencyCovariance.dataframe_to_covariancematrix — Function

dataframe_to_covariancematrix(dd::DataFrame)

Convert a CovarianceMatrix to a DataFrame format.

Inputs

dd - A DataFrame of the form generated by the to_dataframe function.

Returns

A CovarianceMatrix struct.

Internal Functions potentially of use for advanced users

Metrics for distances between CovarianceMatrix structs

HighFrequencyCovariance.calculate_mean_abs_distance — Function

calculate_mean_abs_distance(cov1::CovarianceMatrix, cov2::CovarianceMatrix)

Calculates the mean absolute distance (elementwise in L1 norm) between two CovarianceMatrixs. Undefined if any labels differ between the two CovarianceMatrixs.

Inputs

cov1 - The first CovarianceMatrix

cov2 - The second CovarianceMatrix

Returns

An Tuple with the distance for correlations in first entry and distance for volatilities in the second.

calculate_mean_abs_distance(d1::Dict{Symbol,<:Real}, d2::Dict{Symbol,<:Real})

Calculates the mean absolute distance (elementwise in L1 norm) between two CovarianceMatrixs.

Inputs

d1 - The first Dict

d2 - The second Dict

Returns

A scalar with the mean distance between matching elements.

calculate_mean_abs_distance(d1::Tuple{Dict{Symbol,<:Real},Dict{Symbol,<:Real}}, d2::Tuple{Dict{Symbol,<:Real},Dict{Symbol,<:Real}})

Calculates the distance between the first element of d1 and first of d2. And second elements of each. Returns a Tuple with two scalars for the distances between each element and the corresponding element of the other.

Inputs

d1 - The first Tuple

d2 - The second Tuple

Returns

A Tuple with two scalars representing the distances between Tuple elements.

HighFrequencyCovariance.squared_frobenius — Function

squared_frobenius(x1::AbstractMatrix)

Rearrange the squared frobenius norm of a matrix. Returns a real.

Inputs

x1 The matrix that you want the squared frobenius norm for.

Returns

A Scalar.

HighFrequencyCovariance.squared_frobenius_distance — Function

squared_frobenius_distance(x1::AbstractMatrix, x2::AbstractMatrix = x1)

Rearrange the squared frobenius distance between two matrices. Returns a real.

Inputs

x1 The first matrix.

x2 The second matrix.

Returns

A Scalar.

Used in volatility estimation techniques

HighFrequencyCovariance.default_num_grids — Function

default_num_grids(ts::SortedDataFrame)

This gives a default number of intervals to divide a series of ticks over for the purposes of estimating returns and volatility.

Inputs

ts - The tick data.

Returns

An integer for the number of intervals.

Used in covariance estimation techniques

HighFrequencyCovariance.get_all_refresh_times — Function

get_all_refresh_times(ts::SortedDataFrame, assets::Vector{Symbol} = get_assets(ts); start_time::R = minimum(ts.df[:,ts.time])) where R<:Real

Get a vector of all refresh times when all assets have an updated price. So if there are assets A and B that trade at times (1,5,6,7,10) and (2,5,7,9) then the refresh times are (2,5,7,10) as at these four times there are updated prices for all assets that have happened since the previous refresh time.

Inputs

ts - The tick data.

assets - The assets of interest.

start_time - From what time should we start looking for updated prices.

Returns

A Vector of refresh times.

HighFrequencyCovariance.latest_value — Function

latest_value(ts::SortedDataFrame, at_times::Vector{R}; assets::Vector{Symbol} = get_assets(ts)) where R<:Real

Get the latest price at a each input time.

Inputs

ts - The tick data.

at_times - The times you want the latest prices for.

assets - The assets you want latest prices for.

Returns

A DataFrame. Rows are for each time specified in at_times. Columns are for each asset.

HighFrequencyCovariance.time_between_refreshes — Function

time_between_refreshes(ts::SortedDataFrame; assets::Vector{Symbol} = get_assets(ts))

Get a DataFrame showing how many time is between each refresh and how many ticks in total.

Inputs

ts - Tick data.

assets - A Vector of labels.

Returns

A DataFrame summarising the average number of time between ticks for each asset.

HighFrequencyCovariance.random_value_in_interval — Function

random_value_in_interval(ts::SortedDataFrame, at_times::Vector{R}; assets::Vector{Symbol} = get_assets(ts), twister_arb_value_in_interval::MersenneTwister = MersenneTwister(2604)) where R<:Real

Get a random value in an interval. So if you input times 1,7,8 then for the second entry it will pick a random update (if any exist) between times 1 and 7.

Inputs

ts - The tick data.

at_times - The times that seperate the intervals of interest.

assets - The assets of interest.

twister_arb_value_in_interval - The RNG used in selecting the random interval.

Returns

A DataFrame with prices for each asset from random ticks in each interval.

Kernels used in the BNHLS method

HighFrequencyCovariance.HFC_Kernel — Type

A kernel used in the bnhls covariance method.

HighFrequencyCovariance.parzen — Constant

A parzen kernel used in the bnhls covariance method.

HighFrequencyCovariance.quadratic_spectral — Constant

A quadratic_spectral kernel used in the bnhls covariance method.

HighFrequencyCovariance.fejer — Constant

A fejer kernel used in the bnhls covariance method.

HighFrequencyCovariance.tukey_hanning — Constant

A tukey_hanning kernel used in the bnhls covariance method.

HighFrequencyCovariance.bnhls_2008 — Constant

A bnhls_2008 kernel used in the bnhls covariance method.

Used in nearest correlation regularisation

HighFrequencyCovariance.project_to_S — Function

project_to_S(A::Hermitian, W_root::Union{Hermitian,Diagonal}; W_inv_sqrt::Union{Hermitian,Diagonal} = sqrt_psd(inv(W_root^2)))
project_to_S(A::Diagonal, W_root::Union{Hermitian,Diagonal}; W_inv_sqrt::Union{Hermitian,Diagonal,Missing} = missing)

This maps a matrix to the nearest psd matrix. Wroot should be the principal square root of a psd Hermitian weighting matrix, W. `Winvsqrtshould be the corresponding square root of the inverse of W.nearestpsd_matrix` is a simpler interface for this function however it does not allow weighting matrices to be specified.

Inputs

A - The matrix you want to project to the S space. This can be a Diagonal or a Hermitian. Note that if you input a Diagonal matrix then it is already in the S space and so it will be returned without any calculation.

W_root - The inverse weighting matrix.

W_inv_sqrt - The root of W_root. This is calculated if you don't have it but it can save some calculation effort if you already have it.

Outputs

A Hermitian.

References

Higham, N. J. 2001. Theorem 3.2

HighFrequencyCovariance.project_to_U — Function

project_to_U(A::Union{Diagonal,Hermitian}, invW::Hermitian)

This maps the Hermitian/Hermitian matrix A to the nearest matrix in the U space (the space of all unit diagonal matrices as defined by Higham 2001). The inverse weight matrix invW determines how much to adjust each element to get it to be unit diagonal. In other words it is used to determine what is the nearest correlation matrix. The weight matrix must be Hermitian positive definite. We use the W-norm (as defined by Higham 2001).

Inputs

A - The matrix you want to project to the U space

invW - The inverse weighting matrix.

Outputs

A Diagonal or a Hermitian.

References

Higham, N. J. 2001. Bottom of page 335.

HighFrequencyCovariance.iterate_higham — Function

iterate_higham(Y::Union{Hermitian,Diagonal}, Dykstra::Union{Hermitian,Diagonal}, W_root::Union{Hermitian,Diagonal}, W_inv::Union{Hermitian,Diagonal}, W_inv_sqrt::Union{Hermitian,Diagonal})

Do one iterate mapping the input matrix to the S space (of psd matrices) and then to the U space (unit diagonal and all other entries below 1 in absolute value). Returns the updated matrix and the next iterate's Dykstra correction.

Inputs

Y - The matrix you want to project to the iterate towards the space of valid correlation matrices.

Dykstra - The Dykstra correction matrix.

W_root - The root of W.

W_inv - The inverse of W.

W_inv_sqrt - The root of the inverse of W.

Outputs

A Hermitian.

An updated Dykstra correction matrix.

References

Higham, N. J. 2001. Algorithm 3.3