Helper Functions

StochasticIntegrals.ItoSet

HighFrequencyCovariance.blockwise_estimation

HighFrequencyCovariance.combine_covariance_matrices

HighFrequencyCovariance.construct_matrix_from_eigen

HighFrequencyCovariance.cor2cov

HighFrequencyCovariance.cov2cor

HighFrequencyCovariance.cov2cor_and_vol

HighFrequencyCovariance.covariance

HighFrequencyCovariance.dataframe_to_covariancematrix

HighFrequencyCovariance.duration

HighFrequencyCovariance.generate_random_path

HighFrequencyCovariance.get_assets

HighFrequencyCovariance.get_correlation

HighFrequencyCovariance.get_returns

HighFrequencyCovariance.get_volatility

HighFrequencyCovariance.is_psd_matrix

HighFrequencyCovariance.make_nan_covariance_matrix

HighFrequencyCovariance.put_assets_into_blocks_by_trading_frequency

HighFrequencyCovariance.rearrange

HighFrequencyCovariance.subset_to_tick

HighFrequencyCovariance.subset_to_time

HighFrequencyCovariance.ticks_per_asset

HighFrequencyCovariance.valid_correlation_matrix

StochasticIntegrals.to_dataframe

Working with SortedDataFrame structs

HighFrequencyCovariance.get_assets — Function

get_assets(ts::SortedDataFrame, obs_to_include::Integer = 10)

This returns a vector of all of the assets in the SortedDataFrame with at least some number of observations (10 by default).

Inputs

ts - The tick data.

obs_to_include - An integer for the minimum number of ticks in ts needed for the function to include that asset.

Returns

A Vector{Symbol} with each asset.

HighFrequencyCovariance.ticks_per_asset — Function

ticks_per_asset(ts::SortedDataFrame, assets::Vector{Symbol} = get_assets(ts))

Count the number of observations for each asset.

Inputs

ts - The tick data

assets - A vector with asset Symbols.

Returns

A Dict with the number of observations for each input asset.

HighFrequencyCovariance.duration — Function

duration(ts::SortedDataFrame)

The time elapsed between the first and the last tick in a SortedDataFrame.

Inputs

ts - Tick data.

Returns

A scalar representing this duration.

HighFrequencyCovariance.subset_to_tick — Function

subset_to_tick(ts::SortedDataFrame, n::Integer)

This subsets a SortedDataFrame to only the first n ticks.

Inputs

ts - Tick data.

n - How many ticks to subset to.

Returns

A (smaller) SortedDataFrame.

HighFrequencyCovariance.subset_to_time — Function

subset_to_time(ts::SortedDataFrame, totime::Real)

This subsets a SortedDataFrame to only the first observations up until some time.

Inputs

ts - Tick data.

totime - Up to what time.

Returns

A (smaller) SortedDataFrame.

Working with CovarianceMatrix structs

HighFrequencyCovariance.covariance — Function

covariance(cm::CovarianceMatrix, duration::Real)

This makes a Hermitian matrix for the covariance matrix over some duration.

Inputs

cm - A CovarianceMatrix struct.

duration - A duration. This should be in same units as used in estimating cm's volatilities.

Returns

A Hermitian.

HighFrequencyCovariance.get_correlation — Function

get_correlation(covar::CovarianceMatrix, asset1::Symbol, asset2::Symbol)

Extract the correlation between two assets stored in a CovarianceMatrix.

Inputs

covar - A CovarianceMatrix

asset1 - A Symbol representing an asset.

asset2 - A Symbol representing an asset.

Returns

A Scalar (the correlation coefficient).

HighFrequencyCovariance.get_volatility — Function

get_volatility(covar::CovarianceMatrix, asset1::Symbol)

Get the volatility for a stock from a CovarianceMatrix.

Inputs

covar - A CovarianceMatrix

asset1 - A Symbol representing an asset.

Returns

A Scalar (the volatility).

HighFrequencyCovariance.make_nan_covariance_matrix — Function

make_nan_covariance_matrix(labels::Vector{Symbol})

This makes an empty CovarianceMatrix struct with all volatilities and correlations being NaNs.

Inputs

labels - The names of the asset names for this (empty) CovarianceMatrix.

Returns

An (empty) CovarianceMatrix

HighFrequencyCovariance.combine_covariance_matrices — Function

combine_covariance_matrices(vect::Vector{CovarianceMatrix{T}},
                            cor_weights::Vector{<:Real} = repeat([1.0], length(vect)),
                            vol_weights::Vector{<:Real} = cor_weights) where T<:Real

Combines a vector of CovarianceMatrix structs into one CovarianceMatrix struct.

Inputs

vect - A vector of CovarianceMatrix structs.

cor_weights - A vector for how much to weight the correlations from each covariance matrix (by default they will be equalweighted).

vol_weights - A vector for how much to weight the volatilities from each covariance matrix (by default they will be equalweighted).

Returns

A matrix and a vector of labels for each row/column of the matrix.

HighFrequencyCovariance.rearrange — Function

rearrange(cm::CovarianceMatrix, labels::Vector{Symbol})

Rearrange the order of labels in a CovarianceMatrix.

Takes

cm - A CovarianceMatrix.

labels - A Vector of labels.

Returns

A CovarianceMatrix.

HighFrequencyCovariance.cov2cor — Function

cov2cor(mat::AbstractMatrix)

Converts a matrix (representing a covariance matrix) into a Hermitian correlation matrix and a vector of standard deviations.

Inputs

cor - A matrix.

Returns

A Hermitian.

A Vector of standard deviations (not volatilities).

HighFrequencyCovariance.cor2cov — Function

cor2cov(cor::AbstractMatrix,sdevs::Vector{<:Real})

Converts a correlation matrix and some standard deviations into a Hermitian covariance matrix.

Inputs

cor - A correlation matrix.

sdevs - A vector of standard deviations (not volatilities - use sdevs = sqrt(duration) .* volatilities to convert if necessary).

Returns

A Hermitian.

HighFrequencyCovariance.cov2cor_and_vol — Function

cov2cor_and_vol(mat::AbstractMatrix, duration::Real)

Converts a matrix (representing a covariance matrix) into a Hermitian correlation matrix and a vector of volatilities.

Inputs

cor - A correlation matrix.

duration - The duration that the covariance matrix is for.

Returns

A Hermitian.

A Vector of volatilities.

HighFrequencyCovariance.construct_matrix_from_eigen — Function

construct_matrix_from_eigen(eigenvalues::Vector{<:Real},
                            eigenvectors::Matrix{<:Real})

Constructs a matrix from its eigenvalue decomposition.

Inputs

eigenvalues - A vector of eigenvalues.

eigenvectors - A matrix of eigenvectors. The i'th column corresponds to the i'th eigenvalue.

Returns

A Matrix.

HighFrequencyCovariance.get_returns — Function

get_returns(dd::DataFrame; rescale_for_duration::Bool = false)

Converts a long format DataFrame of prices into a DataFrame of returns.

Inputs

dd - A DataFrame with a column called :Time and all other columns being asset prices in each period.

rescale_for_duration - Should returns be rescaled to reflect a common time interval.

Returns

A DataFrame of returns.

HighFrequencyCovariance.valid_correlation_matrix — Function

valid_correlation_matrix(mat::Hermitian)
valid_correlation_matrix(covar::CovarianceMatrix)

Test if a Hermitian matrix is a valid correlation matrix. This is done by testing if it is psd, if it has a unit diagonal and if all other elements are less than one. If a Hermitian is input then it will be tested. If a CovarianceMatrix is input then its correlation matrix will be tested.

Inputs

A Hermitian matrix or a CovarianceMatrix

Returns

A Bool that is true if mat is a valid correlation matrix and false if not.

HighFrequencyCovariance.is_psd_matrix — Function

is_psd_matrix(mat::Hermitian)
is_psd_matrix(covar::CovarianceMatrix)

Test if a matrix is psd (Positive Semi-Definite). This is done by seeing if all eigenvalues are positive. If a Hermitian is input then it will be tested. If a CovarianceMatrix is input then its correlation matrix will be tested.

Inputs

A Hermitian matrix or a CovarianceMatrix

Returns

A Bool that is true if mat is psd and false if not.

Blocking and Regularisation Functions

HighFrequencyCovariance.put_assets_into_blocks_by_trading_frequency — Function

put_assets_into_blocks_by_trading_frequency(ts::SortedDataFrame,
                                            obs_multiple_for_new_block::Real, func::Symbol,
                                            optional_parameters::NamedTuple = NamedTuple())

This makes a DataFrame that describes how to estimate the covariance matrix blockwise.

Inputs

ts - The tick data.

obs_multiple_for_new_block - The relative number of ticks needed before a new block is made. So if this is 1.2 that means a new group is made when one asset has 20% or more ticks than the slowest traded asset in the previous block.

func - A symbol representing the covariance estimation function to be used.

optional_parameters - Optional parameters to be used in the func function.

Returns

A DataFrame representing what estimations should be performed. The order of rows in the DataFrame shows the order of estimation.

References

Hautsch, N., Kyj, L.M. and Oomen, R.C.A. (2012), A blocking and regularization approach to high‐dimensional realized covariance estimation. J. Appl. Econ., 27: 625-645

HighFrequencyCovariance.blockwise_estimation — Function

blockwise_estimation(ts::SortedDataFrame, blocking_frame::DataFrame)

Run a series of covariance estimations and combine the results. Two things should be input, a SortedDataFrame with the price update data and a dataframe describing what estimations should be performed. This should be of the same form as is output by put_assets_into_blocks_by_trading_frequency (although the actual estimations can be customised to something different as to what that function outputs).

Inputs

ts - The tick data.

blocking_frame - A DataFrame representing what estimations to do and in what order. This is often be one generated by the put_assets_into_blocks_by_trading_frequency function (and potentially then modified).

Returns

A CovarianceMatrix.

Monte Carlo

HighFrequencyCovariance.generate_random_path — Function

generate_random_path(dimensions::Integer, ticks::Integer; syncronous::Bool = false
                     twister::MersenneTwister = MersenneTwister(1), minvol::Real = 0.0,
                     maxvol::Real = 0.02, min_refresh_rate::Real = 1.0,
                     max_refresh_rate::Real = 5.0, min_noise_var::Real = 0.0,
                     max_noise_var::Real = 0.01, assets::Union{Vector,Missing} = missing,
                     brownian_corr_matrix::Union{Hermitian,Missing} = missing,
                     vols::Union{Vector,Missing} = missing)

Generate a random path of price updates with a specified number of dimensions and ticks. There are options for whether the data is syncronous or asyncronous, the volatility of the price processes, the refresh rate on the (exponential) arrival times of price updates, the minimum and the maximum microstructure noises.

Inputs

dimensions - The number of assets.

ticks - The number of ticks to produce.

syncronous - Should ticks be syncronous (for each asset) or asyncronous.

twister - The MersenneTwister used for RNG.

minvol - The minimum volatility in sampling (only used if vols is missing).

maxvol - The maximum volatility in sampling (only used if vols is missing).

min_refresh_rate - The minimum refresh rate in sampling.

max_refresh_rate - The maximum refresh rate in sampling.

min_noise_var - The minimum assetwise microstructure noise variance.

max_noise_var - The minimum assetwise microstructure noise variance.

assets - The names of the assets that you want to use. The length of this must be equal to the dimensions input.

brownian_corr_matrix - The correlation matrix to use. This is sampled from the Inverse Wishart distribution if none is input.

vols - The correlation matrix to use. This is sampled from the Inverse Wishart distribution if none is input.

Returns

A SortedDataFrame of tick data.

A CovarianceMatrix representing the true data generation process used in making the tick data.

A Dict of microstructure noise variances for each asset.

A Dict of update rates for each asset.

StochasticIntegrals.ItoSet — Type

ItoSet(covariance_matrix::CovarianceMatrix{<:Real})

Convert a CovarianceMatrix into an ItoSet from the StochasticIntegrals package. This package can then be used to do things like generate draws from the Multivariate Gaussian corresponding to the covariance matrix and other things.

Inputs

covariance_matrix - The CovarianceMatrix that you want to convert into an StochasticIntegrals.ItoSet

Returns

A StochasticIntegrals.ItoSet struct.

For getting a DataFrame version of a CovarianceMatrix and vice versa.

StochasticIntegrals.to_dataframe — Function

to_dataframe(covar::CovarianceMatrix, othercols::Dict = Dict{Symbol,Any}();
             delete_duplicate_correlations::Bool = true)

Convert a CovarianceMatrix to a DataFrame format.

Inputs

covar - The CovarianceMatrix

othercols - This adds columns to the DataFrame. For instance if it is Dict{Symbol,String}([:pc] .=> ["Fred's PC"]), then there will be a column indicating that this estimation was done on Fred's PC.

delete_duplicate_correlations - Should the unnecessary correlations be included (as correlation matrices are symmetric half the entries duplicate information).

Returns

A DataFrame.

HighFrequencyCovariance.dataframe_to_covariancematrix — Function

dataframe_to_covariancematrix(dd::DataFrame)

Convert a CovarianceMatrix to a DataFrame format.

Inputs

dd - A DataFrame of the form generated by the to_dataframe function.

Returns

A CovarianceMatrix struct.