Internal Functions

HighFrequencyCovariance.bnhls_2008

HighFrequencyCovariance.fejer

HighFrequencyCovariance.parzen

HighFrequencyCovariance.quadratic_spectral

HighFrequencyCovariance.tukey_hanning

HighFrequencyCovariance.HFC_Kernel

HighFrequencyCovariance.calculate_mean_abs_distance

HighFrequencyCovariance.default_num_grids

HighFrequencyCovariance.get_all_refresh_times

HighFrequencyCovariance.iterate_higham

HighFrequencyCovariance.latest_value

HighFrequencyCovariance.project_to_S

HighFrequencyCovariance.project_to_U

HighFrequencyCovariance.random_value_in_interval

HighFrequencyCovariance.squared_frobenius

HighFrequencyCovariance.squared_frobenius_distance

HighFrequencyCovariance.time_between_refreshes

Metrics for distances between CovarianceMatrix structs

HighFrequencyCovariance.calculate_mean_abs_distance — Function

calculate_mean_abs_distance(cov1::CovarianceMatrix, cov2::CovarianceMatrix, decimal_places::Integer = 8; return_nans_if_symbols_dont_match::Bool = true)

Calculates the mean absolute distance (elementwise in L1 norm) between two CovarianceMatrixs. Undefined if any labels differ between the two CovarianceMatrixs.

Inputs

cov1 - The first CovarianceMatrix

cov2 - The second CovarianceMatrix

decimal_places - How many decimal places to show the result to.

return_nans_if_symbols_dont_match - If the symbols don't match should it be an error. Or if false we only compare common symbols in both CovarianceMatrixs

Returns

An Tuple with the distance for correlations in first entry and distance for volatilities in the second.

calculate_mean_abs_distance(d1::Dict{Symbol,<:Real}, d2::Dict{Symbol,<:Real})

Calculates the mean absolute distance (elementwise in L1 norm) between two CovarianceMatrixs.

Inputs

d1 - The first Dict

d2 - The second Dict

Returns

A scalar with the mean distance between matching elements.

HighFrequencyCovariance.squared_frobenius — Function

squared_frobenius(x1::AbstractMatrix)

Returns the squared frobenius norm of a matrix. This is a real.

Inputs

x1 The matrix that you want the squared frobenius norm for.

Returns

A Scalar.

HighFrequencyCovariance.squared_frobenius_distance — Function

squared_frobenius_distance(x1::AbstractMatrix, x2::AbstractMatrix = x1)

Returns the squared frobenius distance between two matrices. This is a real.

Inputs

x1 The first matrix.

x2 The second matrix.

Returns

A Scalar.

Used in volatility estimation techniques

HighFrequencyCovariance.default_num_grids — Function

default_num_grids(ts::SortedDataFrame)

This gives a default number of intervals to divide a series of ticks over for the purposes of estimating returns and volatility.

Inputs

ts - The tick data.

Returns

An integer for the number of intervals.

Used in covariance estimation techniques

HighFrequencyCovariance.get_all_refresh_times — Function

get_all_refresh_times(ts::SortedDataFrame, assets::Vector{Symbol} = get_assets(ts);
                      start_time::R = minimum(ts.df[:,ts.time])) where R<:Real

Get a vector of all refresh times when all assets have an updated price. So if there are assets A and B that trade at times (1,5,6,7,10) and (2,5,7,9) then the refresh times are (2,5,7,10) as at these four times there are updated prices for all assets that have happened since the previous refresh time.

Inputs

ts - The tick data.

assets - The assets of interest.

start_time - From what time should we start looking for updated prices.

Returns

A Vector of refresh times.

HighFrequencyCovariance.latest_value — Function

latest_value(ts::SortedDataFrame, at_times::Vector{<:Real};
             assets::Vector{Symbol} = get_assets(ts))

Get the latest price at a each input time.

Inputs

ts - The tick data.

at_times - The times you want the latest prices for.

assets - The assets you want latest prices for.

Returns

A DataFrame. Rows are for each time specified in at_times. Columns are for each asset.

HighFrequencyCovariance.time_between_refreshes — Function

time_between_refreshes(ts::SortedDataFrame;
                       assets::Vector{Symbol} = get_assets(ts))

Get a DataFrame showing how many time is between each refresh and how many ticks in total.

Inputs

ts - Tick data.

assets - A Vector of labels.

Returns

A DataFrame summarising the average number of time between ticks for each asset.

HighFrequencyCovariance.random_value_in_interval — Function

random_value_in_interval(ts::SortedDataFrame, at_times::Vector{<:Real};
                         assets::Vector{Symbol} = get_assets(ts),
                         twister_arb_value_in_interval::MersenneTwister = MersenneTwister(2604))

Get a random value in an interval. So if you input times 1,7,8 then for the second entry it will pick a random update (if any exist) between times 1 and 7.

Inputs

ts - The tick data.

at_times - The times that seperate the intervals of interest.

assets - The assets of interest.

twister_arb_value_in_interval - The RNG used in selecting the random interval.

Returns

A DataFrame with prices for each asset from random ticks in each interval.

Kernels used in the BNHLS method

HighFrequencyCovariance.HFC_Kernel — Type

A kernel used in the bnhls covariance method.

HighFrequencyCovariance.parzen — Constant

A parzen kernel used in the bnhls covariance method.

HighFrequencyCovariance.quadratic_spectral — Constant

A quadratic_spectral kernel used in the bnhls covariance method.

HighFrequencyCovariance.fejer — Constant

A fejer kernel used in the bnhls covariance method.

HighFrequencyCovariance.tukey_hanning — Constant

A tukey_hanning kernel used in the bnhls covariance method.

HighFrequencyCovariance.bnhls_2008 — Constant

A bnhls_2008 kernel used in the bnhls covariance method.

Used in nearest correlation regularisation

HighFrequencyCovariance.project_to_S — Function

project_to_S(A::Hermitian, W_root::Union{Hermitian,Diagonal};
             W_inv_sqrt::Union{Hermitian,Diagonal} = sqrt_psd(inv(W_root^2)))
project_to_S(A::Diagonal, W_root::Union{Hermitian,Diagonal};
             W_inv_sqrt::Union{Hermitian,Diagonal,Missing} = missing)

This maps a matrix to the nearest psd matrix. W_root should be the principal square root of a psd Hermitian weighting matrix, W. W_inv_sqrt should be the corresponding square root of the inverse of W. nearest_psd_matrix is a simpler interface for this function however it does not allow weighting matrices to be specified.

Inputs

A - The matrix you want to project to the S space. This can be a Diagonal or a Hermitian. Note that if you input a Diagonal matrix then it is already in the S space and so it will be returned without any calculation.

W_root - The inverse weighting matrix.

W_inv_sqrt - The root of W_root. This is calculated if you don't have it but it can save some calculation effort if you already have it.

Outputs

A Hermitian.

References

Higham, N. J. 2001. Theorem 3.2

HighFrequencyCovariance.project_to_U — Function

project_to_U(A::Union{Diagonal,Hermitian}, invW::Hermitian)

This maps the Hermitian/Hermitian matrix A to the nearest matrix in the U space (the space of all unit diagonal matrices as defined by Higham 2001). The inverse weight matrix invW determines how much to adjust each element to get it to be unit diagonal. In other words it is used to determine what is the nearest correlation matrix. The weight matrix must be Hermitian positive definite. We use the W-norm (as defined by Higham 2001).

Inputs

A - The matrix you want to project to the U space

invW - The inverse weighting matrix.

Outputs

A Diagonal or a Hermitian.

References

Higham, N. J. 2001. Bottom of page 335.

HighFrequencyCovariance.iterate_higham — Function

iterate_higham(Y::Union{Hermitian,Diagonal}, Dykstra::Union{Hermitian,Diagonal},
               W_root::Union{Hermitian,Diagonal}, W_inv::Union{Hermitian,Diagonal},
               W_inv_sqrt::Union{Hermitian,Diagonal})

Do one iterate mapping the input matrix to the S space (of psd matrices) and then to the U space (unit diagonal and all other entries below 1 in absolute value). Returns the updated matrix and the next iterate's Dykstra correction.

Inputs

Y - The matrix you want to project to the iterate towards the space of valid correlation matrices.

Dykstra - The Dykstra correction matrix.

W_root - The root of W.

W_inv - The inverse of W.

W_inv_sqrt - The root of the inverse of W.

Outputs

A Hermitian.

An updated Dykstra correction matrix.

References

Higham, N. J. 2001. Algorithm 3.3